refactor: validation-helpers removing
