+++ packages/core-transaction-pool/lib/guard.js
@@ -2,7 +2,7 @@ const Promise = require('bluebird')
 const container = require('@arkecosystem/core-container')
 const { Transaction } = require('@arkecosystem/crypto').models
 const { TRANSACTION_TYPES } = require('@arkecosystem/crypto').constants
-const helpers = require('./utils/validation-helpers')
+const isRecipientOnActiveNetwork = require('./utils/is-on-active-network')
 const database = container.resolvePlugin('database')
 const _ = require('lodash')
 
@@ -151,7 +151,7 @@ module.exports = class TransactionGuard {
   async __determineValidTransactions () {
     await Promise.each(this.transactions, async (transaction) => {
       if (transaction.type === TRANSACTION_TYPES.TRANSFER) {
-        if (!helpers.isRecipientOnActiveNetwork(transaction)) {
+        if (isRecipientOnActiveNetwork(transaction)) {
           this.invalid.push(transaction)
 
           return
