--- packages/core-transaction-pool/lib/utils/validation-helpers.js
@@ -1,24 +1,0 @@
-'use strict'
-const container = require('@arkecosystem/core-container')
-const bs58check = require('bs58check')
-const { configManager } = require('@arkecosystem/crypto')
-const logger = container.resolvePlugin('logger')
-
-module.exports = {
-  /**
-   * Checks if transaction recipient is on the same network as blockchain
-   * @param {Transaction}
-   * @return {Boolean}
-   */
-  isRecipientOnActiveNetwork: transaction => {
-    const recipientPrefix = bs58check.decode(transaction.recipientId).readUInt8(0)
-
-    if (recipientPrefix === configManager.get('pubKeyHash')) {
-      return true
-    }
-
-    logger.error(`Recipient ${transaction.recipientId} in not on the same network: ${configManager.get('pubKeyHash')}`)
-
-    return false
-  }
-}
